+++ tests/integration/test_store.py
@@ -460,51 +460,6 @@ def test_processing(
     assert event.get("version") is not None
 
 
-def test_processing_quotas_legacy(
-    mini_sentry, relay_with_processing, outcomes_consumer, events_consumer
-):
-    relay = relay_with_processing({"processing": {"max_rate_limit": 120}})
-    relay.wait_relay_healthcheck()
-
-    mini_sentry.project_configs[42] = projectconfig = mini_sentry.full_project_config()
-    public_keys = projectconfig["publicKeys"]
-    key_id = public_keys[0]["numericId"]
-    public_keys[0]["quotas"] = [
-        {
-            "prefix": "test_rate_limiting_{}".format(uuid.uuid4().hex),
-            "limit": 5,
-            "window": 3600,
-            "reasonCode": "get_lost",
-        }
-    ]
-
-    events_consumer = events_consumer()
-    outcomes_consumer = outcomes_consumer()
-
-    for i in range(5):
-        relay.send_event(42, {"message": f"regular{i}"})
-
-        event, _ = events_consumer.get_event()
-        assert event["logentry"]["formatted"] == f"regular{i}"
-
-    # this one will not get a 429 but still get rate limited (silently) because
-    # of our caching
-    relay.send_event(42, {"message": "some_message"})
-
-    outcomes_consumer.assert_rate_limited("get_lost", key_id=key_id)
-
-    for _ in range(5):
-        with pytest.raises(HTTPError) as excinfo:
-            relay.send_event(42, {"message": "rate_limited"})
-
-        # The rate limit is actually for 1 hour, but we cap at 120s with the
-        # max_rate_limit parameter
-        retry_after = excinfo.value.response.headers["retry-after"]
-        assert int(retry_after) <= 120
-
-        outcomes_consumer.assert_rate_limited("get_lost", key_id=key_id)
-
-
 def test_processing_quotas(
     mini_sentry, relay_with_processing, outcomes_consumer, events_consumer
 ):
@@ -530,7 +485,6 @@ def test_processing_quotas(
         "publicKey": "31a5a894b4524f74a9a8d0e27e21ba92",
         "isEnabled": True,
         "numericId": 1234,
-        "quotas": [],
     }
     public_keys.append(second_key)