+++ CHANGELOG.md
@@ -19,12 +19,14 @@ We have switched to [CalVer](https://calver.org/)! Relay's version is always in 
 
 **Internal**:
 
+- Add support for Outcomes generation in non processing relays.([#592](https://github.com/getsentry/relay/pull/592))
 - Ignore non-Rust folders for faster rebuilding and testing. ([#578](https://github.com/getsentry/relay/pull/578))
 - Invalid session payloads are now logged for SDK debugging. ([#584](https://github.com/getsentry/relay/pull/584), [#591](https://github.com/getsentry/relay/pull/591))
 - Remove unused `rev` from project state. ([#586](https://github.com/getsentry/relay/pull/586))
 - Add an outcome endpoint for trusted Relays. ([#589](https://github.com/getsentry/relay/pull/589))
 - Emit outcomes for event payloads submitted in attachment files. ([#609](https://github.com/getsentry/relay/pull/609))
 - Split envelopes that contain sessions and other items and ingest them independently. ([#610](https://github.com/getsentry/relay/pull/610))
+- Removed support for legacy per-key quotas. ([#616](https://github.com/getsentry/relay/pull/615))
 
 ## 0.5.9